pyproject.toml
[project]
name = "mdocx"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Donald Thompson", email = "witt3rd@witt3rd.com" }
]
dependencies = [
    "python-docx>=1.1.2",
    "mistletoe>=1.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mdocx"]
